+++ src/commandContext.ts
@@ -16,6 +16,7 @@ export enum CommandContext {
     BbyEnvironmentActive = 'atlascode:bbyEnvironmentActive',
     DebugMode = 'atlascode:debugMode',
     UseNewAuthFlow = 'atlascode:useNewAuthFlow',
+    IsEditorFocused = 'atlascode:isEditorFocused',
 }
 
 export function setCommandContext(key: CommandContext | string, value: any) {