+++ src/webviews/components/issue/view-issue-screen/JiraIssuePage.tsx
@@ -252,12 +252,32 @@ export default class JiraIssuePage extends AbstractIssueEditorPage<Emit, Accept,
             }
             case UIType.Worklog: {
                 this.setState({ isSomethingLoading: true, loadingField: field.key });
-                this.postMessage({
-                    action: 'createWorklog',
-                    site: this.state.siteDetails,
-                    worklogData: newValue,
-                    issueKey: this.state.key,
-                });
+
+                if (newValue.action === 'updateWorklog') {
+                    this.postMessage({
+                        action: 'updateWorklog',
+                        site: this.state.siteDetails,
+                        issueKey: this.state.key,
+                        worklogId: newValue.worklogId,
+                        worklogData: newValue.worklogData,
+                    });
+                } else if (newValue.action === 'deleteWorklog') {
+                    this.postMessage({
+                        action: 'deleteWorklog',
+                        site: this.state.siteDetails,
+                        issueKey: this.state.key,
+                        worklogId: newValue.worklogId,
+                        adjustEstimate: newValue.adjustEstimate,
+                        newEstimate: newValue.newEstimate,
+                    });
+                } else {
+                    this.postMessage({
+                        action: 'createWorklog',
+                        site: this.state.siteDetails,
+                        worklogData: newValue,
+                        issueKey: this.state.key,
+                    });
+                }
                 break;
             }
 
@@ -646,6 +666,7 @@ export default class JiraIssuePage extends AbstractIssueEditorPage<Emit, Accept,
                     isAtlaskitEditorEnabled={this.state.isAtlaskitEditorEnabled}
                     onIssueUpdate={this.handleChildIssueUpdate}
                     mentionProvider={this.mentionProvider}
+                    handleEditorFocus={this.handleEditorFocus}
                 />
                 {this.advancedMain()}
                 {this.state.fields['comment'] && (
@@ -670,6 +691,7 @@ export default class JiraIssuePage extends AbstractIssueEditorPage<Emit, Accept,
                             isEditingComment={this.state.isEditingComment}
                             onEditingCommentChange={this.handleCommentEditingChange}
                             mentionProvider={this.mentionProvider}
+                            handleEditorFocus={this.handleEditorFocus}
                         />
                     </div>
                 )}